massive changes: > newest product list in home page > seperate product search component > product search functionality in both profile & product list page > protected routes for edit & delete product

+++ shopswift/product/views.py
@@ -2,13 +2,56 @@ from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
 from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404
+from django.db.models import Q
 
-from .models import Product
+from .models import Product, Category
 from .forms import ProductForm
 
 
 def list(request):
-    return render(request, "product/list.html")
+    products = Product.objects.all().order_by("-created_at")
+    categories = Category.objects.all()
+    url_query = {}
+
+    if len(request.GET) > 0:
+        name = request.GET["product"]
+        category = request.GET["category"]
+        sort = request.GET["sort"]
+        min_price = int(request.GET["min"]) if request.GET["min"] else None
+        max_price = int(request.GET["max"]) if request.GET["max"] else None
+
+        if name:
+            products = Product.objects.filter(Q(name__icontains=name))
+            url_query["name"] = name
+
+        if category != "all":
+            products = products.filter(Q(category__name=category))
+            url_query["category"] = category
+
+        if min_price:
+            products = products.filter(Q(price__gte=min_price))
+            url_query["min"] = min_price
+        if max_price:
+            products = products.filter(Q(price__lte=max_price))
+            url_query["max"] = max_price
+
+        url_query["sort"] = sort
+        if sort == "low":
+            products = products.order_by("price")
+        elif sort == "high":
+            products = products.order_by("-price")
+        elif sort == "order":
+            products = products.order_by("name")
+        elif sort == "reverse":
+            products = products.order_by("-name")
+
+    context = {
+        "products": products,
+        "categories": categories,
+        "url_query": url_query,
+    }
+
+    return render(request, "product/list.html", context)
 
 
 def detail(request, product_id):
@@ -20,6 +63,8 @@ def detail(request, product_id):
 @login_required
 def edit(request, product_id):
     product = get_object_or_404(Product, id=product_id)
+    if request.user != product.created_by:
+        return redirect("product:list")
     if request.method == "POST":
         form = ProductForm(request.POST, request.FILES, instance=product)
         if form.is_valid():
@@ -51,6 +96,8 @@ def create(request):
 @login_required
 def delete(request, product_id):
     product = get_object_or_404(Product, id=product_id)
+    if request.user != product.created_by:
+        return redirect("product:list")
     if product is not None:
         product.delete()
-        return HttpResponseRedirect(request.META.get("HTTP_REFERER"))
+        return redirect("product:list")
